moving shared initializers to global.go
